Extract grid and sidebar components.
